requirements.txt
requests
PyMySQL
suds-jurko